+++ Source/OxyPlot.SkiaSharp/SkiaRenderContext.cs
@@ -46,6 +46,11 @@ namespace OxyPlot.SkiaSharp
         public bool UseTextShaping { get; set; } = true;
 
         /// <summary>
+        /// Gets or sets the Miter limit. This is the maximum ratio between Miter length and stroke thickness. When this ration is exceeded, the join falls back to a Bevel. The default value is 10.
+        /// </summary>
+        public float MiterLimit { get; set; } = 10;
+
+        /// <summary>
         /// Gets a value indicating whether the context renders to pixels.
         /// </summary>
         /// <value><c>true</c> if the context renders to pixels; otherwise, <c>false</c>.</value>
@@ -826,6 +831,7 @@ namespace OxyPlot.SkiaSharp
             this.paint.StrokeWidth = this.GetActualThickness(strokeThickness, edgeRenderingMode);
             this.paint.PathEffect = null;
             this.paint.StrokeJoin = SKStrokeJoin.Miter;
+            this.paint.StrokeMiter = this.MiterLimit;
             return this.paint;
         }